Rename the default karaoke sprite text into drawable karaoke sprite text. Let the wording be more generic.
